+++ apps/style/code-studio/curriculum_catalog_filters.module.scss
@@ -1,0 +1,49 @@
+.catalogFiltersContainer {
+  position: sticky;
+  top: 0;
+  background-color: white;
+}
+
+.catalogDropdownFilters {
+  display: flex;
+  align-items: baseline;
+  gap: 16px;
+  max-width: 60em;
+  margin: 0 auto;
+  padding: 1em 0;
+
+  button {
+    margin: 0;
+  }
+}
+
+.catalogLanguageFilterRow {
+  display: flex;
+  justify-content: space-between;
+  max-width: 60em;
+  margin: 0 auto;
+  padding: 0 0 1em 0;
+
+  p {
+    margin: 1em 0;
+  }
+
+  i {
+    margin: auto 0.25em;
+    font-size: 1.75em;
+  }
+
+  span {
+    font-size: 16px;
+  }
+}
+
+.catalogLanguageFilterRowNumAvailable {
+  display: flex;
+  flex: start;
+}
+
+button.catalogClearFiltersButton {
+  font-size: 16px;
+  margin-left: auto;
+}
